requirements.txt
langchain 
wikipedia 
langchain-community 
openrouter
python-dotenv
pydantic
duckduckgo-search